requirements.txt
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
protobuf==3.19.1
pycparser==2.21
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
yapf==0.31.0
